Added new `timeProvider` property to a logger.
Allows to set an external time provider closure for log messages.
If `nil` then default behavior `Date()` will be used.

This feature allows to "mock" the time - could be useful in unit tests
which may run on some "virtual" time instead of real one.

Child loggers inherit this value so it's possible to set it only once
(on the `rootLogger` for example) - entire sub-hierarchy of loggers will have it.